Move config.resolversNonOptionalTypename
